Reverting back to last version.  Floats made CPU-usage worse in this case.

+++ src/map_shp.c
@@ -1458,8 +1458,8 @@ void draw_shapefile_map (Widget w,
                             // lines will get you into trouble.
                             x = my_long - x_long_offset;
                             y = my_lat - y_lat_offset;
-                            x = (unsigned long)((float)x / (float)scale_x);
-                            y = (unsigned long)((float)y / (float)scale_y);
+                            x = x / scale_x;
+                            y = y / scale_y;
 
                             if (x >  16000) ok = 0;     // Skip this point
                             if (x < -16000) ok = 0;     // Skip this point
@@ -1536,8 +1536,8 @@ void draw_shapefile_map (Widget w,
                             // lines will get you into trouble.
                             x = my_long - x_long_offset;
                             y = my_lat - y_lat_offset;
-                            x = (unsigned long)((float)x / (float)scale_x);
-                            y = (unsigned long)((float)y / (float)scale_y);
+                            x = x / scale_x;
+                            y = y / scale_y;
 
 
                             // Save the endpoints of the first line
@@ -2163,8 +2163,8 @@ void draw_shapefile_map (Widget w,
                             // lines will get you into trouble.
                             x = my_long - x_long_offset;
                             y = my_lat - y_lat_offset;
-                            x = (unsigned long)((float)x / (float)scale_x);
-                            y = (unsigned long)((float)y / (float)scale_y);
+                            x = x / scale_x;
+                            y = y / scale_y;
 
 
                             if (x >  16000) ok = 0;     // Skip this point
@@ -2541,8 +2541,8 @@ void draw_shapefile_map (Widget w,
                             // lines will get you into trouble.
                             x = my_long - x_long_offset;
                             y = my_lat - y_lat_offset;
-                            x = (unsigned long)((float)x / (float)scale_x);
-                            y = (unsigned long)((float)y / (float)scale_y);
+                            x = x / scale_x;
+                            y = y / scale_y;
 
 
                             // Here we check for really wacko points that will cause problems
@@ -2573,8 +2573,8 @@ void draw_shapefile_map (Widget w,
                             // lines will get you into trouble.
                             width = my_long - x_long_offset;
                             height = my_lat - y_lat_offset;
-                            width = (long)((float)width / (float)scale_x);
-                            height = (long)((float)height / (float)scale_y);
+                            width = width / scale_x;
+                            height = height / scale_y;
 
                             // Here we check for really wacko points that will cause problems
                             // with the X drawing routines, and fix them.
@@ -2676,8 +2676,8 @@ void draw_shapefile_map (Widget w,
                                         // lines will get you into trouble.
                                         x = my_long - x_long_offset;
                                         y = my_lat - y_lat_offset;
-                                        x = (unsigned long)((float)x / (float)scale_x);
-                                        y = (unsigned long)((float)y / (float)scale_y);
+                                        x = x / scale_x;
+                                        y = y / scale_y;
 
         
                                         //fprintf(stderr,"%ld %ld\t\t", x, y);
@@ -2865,8 +2865,8 @@ if (on_screen) {
                                 // lines will get you into trouble.
                                 x = my_long - x_long_offset;
                                 y = my_lat - y_lat_offset;
-                                x = (unsigned long)((float)x / (float)scale_x);
-                                y = (unsigned long)((float)y / (float)scale_y);
+                                x = x / scale_x;
+                                y = y / scale_y;
 
 
                                 //fprintf(stderr,"%ld %ld\t\t", x, y);
@@ -3264,8 +3264,8 @@ if (on_screen) {
                             // lines will get you into trouble.
                             x = my_long - x_long_offset;
                             y = my_lat - y_lat_offset;
-                            x = (unsigned long)((float)x / (float)scale_x);
-                            y = (unsigned long)((float)y / (float)scale_y);
+                            x = x / scale_x;
+                            y = y / scale_y;
 
 
                             if (x >  16000) ok = 0;     // Skip this point
